test(ci): advance the sweeper reopen grace virtually
